YoloFSOD/requirements.txt
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
flask
ultralytics
pyyaml
